client/requirements.txt
celery==4.4.7
vine==1.3.0
redis
urllib3
requests
tld
bs4
dnspython
lxml
aiohttp
PyYAML
asyncio